+++ src/main/java/ninja/eivind/hotsreplayuploader/services/UploaderService.java
@@ -20,8 +20,8 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.concurrent.ScheduledService;
 import javafx.concurrent.Task;
+import ninja.eivind.hotsreplayuploader.concurrent.tasks.UploadTask;
 import ninja.eivind.hotsreplayuploader.files.AccountDirectoryWatcher;
-import ninja.eivind.hotsreplayuploader.files.UploadTask;
 import ninja.eivind.hotsreplayuploader.models.ReplayFile;
 import ninja.eivind.hotsreplayuploader.models.Status;
 import ninja.eivind.hotsreplayuploader.repositories.FileRepository;
